Move nim build flags to a new project-side nim.cfg
Previously, the nim compiler build flags were in the Python script.
Since command-line settings take priority over configuration file settings,
this tool would overwrite a user's build configuration.  We should avoid this.

This commit moves the build flags to a new nim.cfg file that
gets copied to the user's project folder.  This allows the user to change
the build settings as needed.
Added the "opt:size" build flag to nim.cfg.

There remain two flags that are set by this tool: path and cpu
because they come from the platformio values in the Python script.